requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cssselect==1.1.0
cycler==0.11.0
Cython==0.29.23
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
fonttools==4.30.0
funcy==1.17
future==0.18.2
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.2
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lime==0.2.0.1
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
newspaper3k==0.2.8
nltk==3.7
notebook==6.4.8
numexpr==2.8.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
pyLDAvis==3.3.1
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.2
requests==2.27.1
requests-file==1.5.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
streamlit==1.7.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.2.9
tinysegmenter==0.3
tldextract==3.2.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wincertstore==0.2
xgboost==1.5.2
zipp==3.7.0
